// src/app/core/data-services/model/getNotificationListResponseDto.ts
/**
 * ENTropy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GetNotificationResponseDto } from './getNotificationResponseDto';


export interface GetNotificationListResponseDto { 
    total: number;
    page: number;
    limit: number;
    items: Array<GetNotificationResponseDto>;
}