// src/app/core/data-services/model/notificationsFindAllDefaultResponse.ts
/**
 * ENTropy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GetNotificationListResponseDto } from './getNotificationListResponseDto';


export interface NotificationsFindAllDefaultResponse { 
    /**
     * Indicates if the operation was successful
     */
    success: boolean;
    /**
     * The error message, if the operation failed
     */
    error?: object | null;
    data?: GetNotificationListResponseDto;
}